+++ src/hal/__init__.py
@@ -21,6 +21,3 @@ __version__ = get_package_version(path=__file__, package_name=NAME)
 
 
 from .actor import HALActor
-
-
-HALCommandType = Command[HALActor]