+++ src/hal/helpers/ffs.py
@@ -10,10 +10,15 @@ from __future__ import annotations
 
 import enum
 
+from typing import TYPE_CHECKING
+
 from hal import config
-from hal.actor import HALCommandType
 
 from . import HALHelper
+
+
+if TYPE_CHECKING:
+    from hal.actor import HALCommandType
 
 
 __all__ = ["FFSHelper"]