+++ src/hal/helpers/lamps.py
@@ -11,10 +11,16 @@ from __future__ import annotations
 import asyncio
 import time
 
-from hal import HALCommandType, config
+from typing import TYPE_CHECKING
+
+from hal import config
 from hal.exceptions import HALError
 
 from . import HALHelper
+
+
+if TYPE_CHECKING:
+    from hal.actor import HALCommandType
 
 
 __all__ = ["LampsHelper"]