+++ src/hal/macros/expose.py
@@ -116,13 +116,14 @@ class ExposeMacro(Macro):
                 )
             )
 
-            tasks.append(
-                self.helpers.apogee.shutter(
-                    self.command,
-                    open=self.config["with_fpi"],
-                    shutter="fpi",
+            if self.config["with_fpi"]:
+                tasks.append(
+                    self.helpers.apogee.shutter(
+                        self.command,
+                        open=True,
+                        shutter="fpi",
+                    )
                 )
-            )
 
         await asyncio.gather(*tasks)